PhloemProject/requirements.txt
pyserial
requests
pyyaml
streamlit
watchdog
pandas